Code split for Plugin domain object
